+++ ng2Client/app/components/content/knowledgeIndex.ts
@@ -1,11 +1,10 @@
 ﻿import { Component, EventEmitter, OnDestroy, OnInit } from "@angular/core";
 import { Observable } from "rxjs/Observable";
 
-import { ChartConfig } from "../../modules/ng-chart/chart-config";
-import { ChartDataItem } from "../../modules/ng-chart/chart-data-item";
+import { ChartConfig, ChartDataItem } from "../../modules/ng-chart/ng-chart.module";
 import { DataService } from "../../services/data.service";
 import { Logger } from "../../services/logger.service";
-import { ResourcePoolService } from "../../services/resource-pool-service";
+import { ResourcePoolService } from "../../services/resource-pool.service";
 import { Settings } from "../../settings/settings";
 
 //declare const __moduleName: string;