+++ ng2Client/app/main-aot.ts
@@ -2,6 +2,7 @@ import { enableProdMode } from "@angular/core";
 import { platformBrowser } from "@angular/platform-browser";
 
 import { AppModuleNgFactory } from "../aot/app/app.module.ngfactory";
+//import { AppModuleNgFactory } from "../aot/app/dev/app.module.ngfactory";
 import { Settings } from "./settings/settings";
 
 // Enable production mode