+++ ng2Client/systemjs.config.js
@@ -16,11 +16,11 @@
             "@angular/core": "npm:@angular/core/bundles/core.umd",
             "@angular/common": "npm:@angular/common/bundles/common.umd",
             "@angular/compiler": "npm:@angular/compiler/bundles/compiler.umd",
+            "@angular/forms": "npm:@angular/forms/bundles/forms.umd",
+            "@angular/http": "npm:@angular/http/bundles/http.umd",
             "@angular/platform-browser": "npm:@angular/platform-browser/bundles/platform-browser.umd",
             "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic/bundles/platform-browser-dynamic.umd",
-            "@angular/http": "npm:@angular/http/bundles/http.umd",
             "@angular/router": "npm:@angular/router/bundles/router.umd",
-            "@angular/forms": "npm:@angular/forms/bundles/forms.umd",
 
             // angular highcharts
             "angular2-highcharts": "npm:angular2-highcharts/index",
@@ -37,7 +37,7 @@
             "angular2-toaster": "npm:angular2-toaster/angular2-toaster",
 
             // breeze
-            "breeze-client": "npm:breeze-client",
+            "breeze-client": "npm:breeze-client/breeze.base.debug",
             "breeze.ajax.angular": "npm:breeze-client/breeze.ajax.angular",
             "breeze-bridge-angular2": "npm:breeze-bridge-angular2/index",
             "breeze.dataService.odata": "npm:breeze-client/breeze.dataService.odata",
@@ -53,8 +53,8 @@
 
             // Traceur (for ES6, which is not there yet)
             "plugin-traceur": "npm:systemjs-plugin-traceur/plugin-traceur.js",
-            "traceur": "npm:systemjs-plugin-traceur/node_modules/traceur/bin/traceur.js",
-            "traceur-runtime": "npm:systemjs-plugin-traceur/node_modules/traceur/bin/traceur-runtime.js"
+            "traceur": "npm:traceur/bin/traceur.js",
+            "traceur-runtime": "npm:traceur/bin/traceur-runtime.js"
         },
         meta: {
             "traceur": {
@@ -68,9 +68,6 @@
         },
         transpiler: "plugin-traceur",
         transpilerRuntime: false,
-        packages: {
-            "breeze-client": { main: "breeze.base.debug" },
-        }
     };
 
     System.config(config);